[fix] improve codes
