+++ flink-table-store-core/src/main/java/org/apache/flink/table/store/table/source/BatchDataTableScan.java
@@ -18,65 +18,22 @@
 
 package org.apache.flink.table.store.table.source;
 
-import org.apache.flink.table.store.CoreOptions;
-import org.apache.flink.table.store.file.utils.SnapshotManager;
 import org.apache.flink.table.store.table.DataTable;
-import org.apache.flink.table.store.table.source.snapshot.SnapshotSplitReader;
 import org.apache.flink.table.store.table.source.snapshot.StartingScanner;
-
-import javax.annotation.Nullable;
 
 import java.io.Serializable;
 
 /** {@link DataTableScan} for batch planning. */
-public abstract class BatchDataTableScan extends AbstractDataTableScan {
+public interface BatchDataTableScan extends DataTableScan {
 
-    private final SnapshotManager snapshotManager;
-
-    private StartingScanner startingScanner;
-
-    private boolean hasNext;
-
-    public BatchDataTableScan(
-            CoreOptions options,
-            SnapshotSplitReader snapshotSplitReader,
-            SnapshotManager snapshotManager) {
-        super(options, snapshotSplitReader);
-        this.snapshotManager = snapshotManager;
-        this.hasNext = true;
-    }
-
-    public BatchDataTableScan withStartingScanner(StartingScanner startingScanner) {
-        this.startingScanner = startingScanner;
-        return this;
-    }
-
-    @Override
-    @Nullable
-    public DataFilePlan plan() {
-        if (startingScanner == null) {
-            startingScanner = createStartingScanner(false);
-        }
-
-        if (hasNext) {
-            hasNext = false;
-            return startingScanner.getPlan(snapshotManager, snapshotSplitReader);
-        } else {
-            return null;
-        }
-    }
-
-    public static BatchDataTableScan create(DataTable dataTable) {
-        BatchDataTableScan scan = (BatchDataTableScan) dataTable.newScan();
-        return scan.withStartingScanner(scan.createStartingScanner(false));
-    }
+    BatchDataTableScan withStartingScanner(StartingScanner startingScanner);
 
     // ------------------------------------------------------------------------
     //  factory interface
     // ------------------------------------------------------------------------
 
     /** Factory to create {@link BatchDataTableScan}. */
-    public interface Factory extends Serializable {
+    interface Factory extends Serializable {
         BatchDataTableScan create(DataTable dataTable);
     }
 }
